Refactoring: - Merged `AbstractDummySwirldState` and `DummySwirldState` - Renamed `DummySwirldState` to `BlockingSwirldState` - Made `BlockingSwirldState` instance of `MerkleRoot` to mimic `MerkleStateRoot`
