+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/StateTests.java
@@ -29,7 +29,7 @@ import com.swirlds.common.test.fixtures.junit.tags.TestComponentTags;
 import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
 import com.swirlds.platform.state.MerkleRoot;
 import com.swirlds.platform.state.State;
-import com.swirlds.platform.test.fixtures.state.DummySwirldState;
+import com.swirlds.platform.test.fixtures.state.BlockingSwirldState;
 import java.io.IOException;
 import java.nio.file.Path;
 import org.junit.jupiter.api.BeforeAll;
@@ -56,7 +56,7 @@ class StateTests {
 
         state = new State();
         state.setPlatformState(randomPlatformState());
-        state.setSwirldState(new DummySwirldState());
+        state.setSwirldState(new BlockingSwirldState());
 
         state.invalidateHash();
         MerkleCryptoFactory.getInstance().digestTreeSync(state);
